backend/requirements.txt
fastapi
uvicorn
azure-cognitiveservices-speech
requests 
numpy
webrtcvad 